requirements.txt
boto3
botocore
certifi
charset-normalizer
idna
jmespath
python-dateutil
python-dotenv
requests
s3transfer
six
urllib3

